Refactor TagEditorView
